gin: Replace gin_comms with std::array in GIN resource
Replace std::unordered_map<uint32_t, ...> with std::array<..., GIN_MAX_COMMS>
for gin_comms, converting get_comm/set_comm to direct O(1) array indexing with
bounds checks. This eliminates hash table overhead on the hot CQ completion
path.

In addition, regroup member variables of nccl_ofi_gin_resources by access
frequency to improve cache locality, including:
- Move gin_ep to the beginning due to frequent lock access.
- Group comm_id_pool and ref_cnt together.
- Place the 8KB gin_comms array at the end to avoid pushing other members apart

+++ src/rdma/gin/nccl_ofi_gin_reqs.cpp
@@ -80,7 +80,7 @@ int nccl_net_ofi_gin_recv_req_t::handle_cq_entry(struct fi_cq_entry *cq_entry_ba
 
 	if (cq_entry->flags & FI_REMOTE_WRITE) {
 		/* RDMA write-immediate completion */
-		uint32_t comm_id = GIN_IMM_GET_COMM_ID(cq_entry->data);
+		uint16_t comm_id = GIN_IMM_GET_COMM_ID(cq_entry->data);
 		auto &gin_comm = resources.get_comm(comm_id);
 
 		ret = gin_comm.handle_signal_write_completion(cq_entry, src_addr, rail_id_arg);
